refactor: separate data in user-info-content
